change the footer and header
since initially i tend to set the footer and header to be white background, but after try some times, i found out it is not so common and it is even so abrupt to use white and blue in footer, since it may catch more attention than the main content. thus i choose the black background. And maybe latter on i will change?